pinc: renamed pin_constrain_loc.h to pinc_main.h
